Move navigation coordination away from the SettingsScreenCoordinator and into the SettingsFlowCoordinator
